+++ pass_4.c
@@ -614,7 +614,7 @@ int pass_4(void) {
             return FAILED;
           }
           else if (tmp_def->type != DEFINITION_TYPE_STACK) {
-            o = tmp_def->value;
+            o = (int)tmp_def->value;
             x = 1;
 
             if (mem_insert(o & 0xFF) == FAILED)
@@ -692,7 +692,7 @@ int pass_4(void) {
             return FAILED;
           }
           else if (tmp_def->type != DEFINITION_TYPE_STACK) {
-            o = tmp_def->value;
+            o = (int)tmp_def->value;
             x = 1;
 
             if (mem_insert(o & 0xFF) == FAILED)
@@ -728,7 +728,7 @@ int pass_4(void) {
             return FAILED;
           }
           else if (tmp_def->type != DEFINITION_TYPE_STACK) {
-            o = tmp_def->value;
+            o = (int)tmp_def->value;
             x = 1;
 
             if (mem_insert(o & 0xFF) == FAILED)
@@ -760,7 +760,7 @@ int pass_4(void) {
             return FAILED;
           }
           else if (tmp_def->type != DEFINITION_TYPE_STACK) {
-            o = tmp_def->value;
+            o = (int)tmp_def->value;
             x = 1;
 
             if (mem_insert(o & 0xFF) == FAILED)
